Note about reverting custom memory values
Unsetting a custom :memory param on a box will not take effect until the box
is destroyed and recreated. This is because omitting the `modifyvm --memory`
argument has no affect on an existing box that has received `--memory` in
the past. The default from the basebox is lost for the lifetime of that box.

We could copy and assert the default in the Vagrantfile but it would risk us
getting out of step with veewee-defs and they may be some performance
penalty in apply it to every up/reload action.